+++ client/src/hooks/useAuth.ts
@@ -1,0 +1,24 @@
+import { useEffect } from 'react';
+import { useAppDispatch, useAppSelector } from './reduxHook';
+import { login } from '../store/slice/loginSlice';
+
+export function useAuth() {
+  const dispatch = useAppDispatch();
+  const isLogin = useAppSelector((state) => state.login.isLogin);
+
+  useEffect(() => {
+    if (!isLogin) {
+      const currentURL = document.location.search;
+      const params = new URLSearchParams(currentURL);
+
+      const accessToken = params.get('access_token');
+      const refreshToken = params.get('refresh_token');
+
+      if (accessToken && refreshToken) {
+        dispatch(login({ accessToken, refreshToken }));
+        localStorage.setItem('accessToken', accessToken);
+        localStorage.setItem('refreshToken', refreshToken);
+      }
+    }
+  }, []);
+}
